+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Features
 
+- Add a read-only Learning to Rank model registry listing ([#943](https://github.com/opensearch-project/dashboards-search-relevance/pull/943))
+- Add a Learning to Rank model upload flow ([#944](https://github.com/opensearch-project/dashboards-search-relevance/pull/944))
+
 ### Enhancements
 
 ### Bug Fixes